backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
marshmallow
python-dotenv
numpy==1.26.4
pytest==8.3.5
pandas
reportlab
werkzeug
openai
gunicorn==20.1.0
